change LikeCommentServlet to apply to any Datastore entity
